+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.2.0
 
 * Add optional item requirements
+* Fix dispensing bottled xp
 
 # 2.1.0